fu-engine: check if plugin changed after update_detach()
If the device is rebooted into a different shape, the plugin managing
the device may also be different, so take that into account.

This would be e.g. the case for plugins that just subclass the
update_detach() method, and leave to other plugins the actual required
firmware update procedure.